K8SPS-95: Add backupSource support for restores (#174)
* K8SPS-95: Add backupSource support for restores

https://jira.percona.com/browse/K8SPS-95

* test fix

* merge fix

* fix shellcheck warnings

* fix `run-restore.sh`

* fix test

* increase timeouts

+++ build/run-restore.sh
@@ -6,44 +6,43 @@ set -o xtrace
 DATADIR=${DATADIR:-/var/lib/mysql}
 PARALLEL=$(grep -c processor /proc/cpuinfo)
 XBCLOUD_ARGS="--curl-retriable-errors=7 --parallel=${PARALLEL}"
-INSECURE_ARG=""
 if [ -n "$VERIFY_TLS" ] && [[ $VERIFY_TLS == "false" ]]; then
-	INSECURE_ARG="--insecure"
+	XBCLOUD_ARGS="${XBCLOUD_ARGS} --insecure"
 fi
 
 run_s3() {
-	xbcloud get ${XBCLOUD_ARGS} ${INSECURE_ARG} ${BACKUP_DEST} --storage=s3 --s3-bucket=${S3_BUCKET}
+	xbcloud get "${XBCLOUD_ARGS}" "${BACKUP_DEST}" --storage=s3 --s3-bucket="${S3_BUCKET}"
 }
 
 run_gcs() {
-	xbcloud get ${XBCLOUD_ARGS} ${INSECURE_ARG} ${BACKUP_DEST} --storage=google --google-bucket="${GCS_BUCKET}"
+	xbcloud get "${XBCLOUD_ARGS}" "${BACKUP_DEST}" --storage=google --google-bucket="${GCS_BUCKET}"
 }
 
 run_azure() {
-	xbcloud get ${XBCLOUD_ARGS} ${INSECURE_ARG} ${BACKUP_DEST} --storage=azure
+	xbcloud get "${XBCLOUD_ARGS}" "${BACKUP_DEST}" --storage=azure
 }
 
 extract() {
 	local targetdir=$1
 
-	xbstream -xv -C ${targetdir} --parallel=${PARALLEL}
+	xbstream -xv -C "${targetdir}" --parallel="${PARALLEL}"
 }
 
 main() {
 	echo "Starting restore ${RESTORE_NAME}"
-	echo "Restoring to backup ${BACKUP_NAME}: ${BACKUP_DEST}"
+	echo "Restoring to backup: ${BACKUP_DEST}"
 
-	rm -rf ${DATADIR}/*
-	tmpdir=$(mktemp --directory ${DATADIR}/${BACKUP_NAME}_XXXX)
+	rm -rf "${DATADIR:?}"/*
+	tmpdir=$(mktemp --directory "${DATADIR}/${RESTORE_NAME}_XXXX")
 
 	case ${STORAGE_TYPE} in
-		"s3") run_s3 | extract ${tmpdir} ;;
-		"gcs") run_gcs | extract ${tmpdir} ;;
-		"azure") run_azure | extract ${tmpdir} ;;
+		"s3") run_s3 | extract "${tmpdir}" ;;
+		"gcs") run_gcs | extract "${tmpdir}" ;;
+		"azure") run_azure | extract "${tmpdir}" ;;
 	esac
 
-	xtrabackup --prepare --rollback-prepared-trx --target-dir=${tmpdir}
-	xtrabackup --datadir=${DATADIR} --move-back --force-non-empty-directories --target-dir=${tmpdir}
+	xtrabackup --prepare --rollback-prepared-trx --target-dir="${tmpdir}"
+	xtrabackup --datadir="${DATADIR}" --move-back --force-non-empty-directories --target-dir="${tmpdir}"
 
 	rm -rf "${tmpdir}"
